Revert "Chưa fix được, tạm thời để như thế"
This reverts commit 79cf67aae941e03be514d43b5053df2905981e0f.

+++ src/ChartPlot.h
@@ -25,7 +25,7 @@ public:
     void resetColor() { nextColorIndex = 0; }
 
     // @trung
-    void changeMaxMinValue(double max, double min);
+    void changeMaxMinValue(int max, int min);
 
 public slots:
